Blenloader API: lib linking and animdata read/write [WIP]
In this patch I continued to experiment with the blenloader api.
Also see D7061 and D7062.

This patch adds a few more direct linking utilities as well as a method for getting the new address of an id data block.
To test the API, I move writing and reading of `AnimData` to `blenkernel`. This includes nla strips, fcurves and id properties.

Differential Revision: https://developer.blender.org/D7321