Asset Catalogs: add catalog ID to AssetMetaData
Add field `char AssetMetaData::catalog_id[MAX_NAME]`, to hold the
catalog ID of the asset. This field should be set using the
`BKE_asset_metadata_catalog_id_set()` function, which ensures that there
are no spaces in the ID and that the field is not too long.

For now only used to show in the UI, for development/debugging purposes.