Assets: add AssetLibrary struct
Add `AssetLibrary` struct that owns a pointer to an `AssetCatalogService`.

The `struct AssetLibrary` is an opaque struct that C code can just pass
around. It is created with `BKE_asset_library_load(library_path)` and
freed with `BKE_asset_library_free(the_pointer)`.

Note that this API is still experimental.