Asset Catalogs: allow adding new catalogs
New catalogs can be added with `AssetCatalogService::create_catalog()`.
It only requires the catalog path, and convert that into a suitable
catalog ID itself. It is intended that artists won't have to bother with
assigning catalog IDs, and that editing those is more of an advanced
feature.

Adding a catalog will automatically save the catalog definition file to
disk.